+++ src/metric_find_query.ts
@@ -16,41 +16,63 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import { chain, map as _map, uniq } from 'lodash';
-import { lastValueFrom } from 'rxjs';
-import { map } from 'rxjs/operators';
+import { chain, map as _map } from 'lodash';
 
-import { MetricFindValue, TimeRange } from '@grafana/data';
+import { getDefaultTimeRange, MetricFindValue, TimeRange } from '@grafana/data';
 
 import { PrometheusDatasource } from './datasource';
-import { getTimeSrv } from './services/TimeSrv';
-import { PromQueryRequest } from './types';
+import { getVictoriaMetricsTime } from './language_utils';
+import {
+  PrometheusLabelNamesRegex,
+  PrometheusLabelNamesRegexWithMatch,
+  PrometheusMetricNamesRegex,
+  PrometheusQueryResultRegex,
+} from './migrations/variableMigration';
 
 export default class PrometheusMetricFindQuery {
   range: TimeRange;
 
-  constructor(private datasource: PrometheusDatasource, private query: string) {
+  constructor(
+    private datasource: PrometheusDatasource,
+    private query: string
+  ) {
     this.datasource = datasource;
     this.query = query;
-    this.range = getTimeSrv().timeRange();
+    this.range = getDefaultTimeRange();
   }
 
-  process(): Promise<MetricFindValue[]> {
-    const labelNamesRegex = /^label_names\(\)\s*$/;
-    const labelValuesRegex = /^label_values\((?:(.+),\s*)?([a-zA-Z_\\.][a-zA-Z0-9_\\.]*)\)\s*$/;
-    const metricNamesRegex = /^metrics\((.+)\)\s*$/;
-    const queryResultRegex = /^query_result\((.+)\)\s*$/;
+  process(timeRange: TimeRange): Promise<MetricFindValue[]> {
+    this.range = timeRange;
+    const labelNamesRegex = PrometheusLabelNamesRegex;
+    const labelNamesRegexWithMatch = PrometheusLabelNamesRegexWithMatch;
+    const labelValuesRegex = /^label_values\((?:(.+),\s*)?([a-zA-Z_][a-zA-Z0-9_]*)\)\s*$/;
+    const metricNamesRegex = PrometheusMetricNamesRegex;
+    const queryResultRegex = PrometheusQueryResultRegex;
     const labelNamesQuery = this.query.match(labelNamesRegex);
+    const labelNamesMatchQuery = this.query.match(labelNamesRegexWithMatch);
+
+    if (labelNamesMatchQuery) {
+      const selector = `{__name__=~".*${labelNamesMatchQuery[1]}.*"}`;
+      return this.datasource.languageProvider.getSeriesLabels(selector, []).then((results) =>
+        results.map((result) => ({
+          text: result,
+        }))
+      );
+    }
+
     if (labelNamesQuery) {
-      return this.labelNamesQuery();
+      return this.datasource.getTagKeys({ filters: [], timeRange });
     }
 
     const labelValuesQuery = this.query.match(labelValuesRegex);
     if (labelValuesQuery) {
-      if (labelValuesQuery[1]) {
-        return this.labelValuesQuery(labelValuesQuery[2], labelValuesQuery[1]);
+      const filter = labelValuesQuery[1];
+      const label = labelValuesQuery[2];
+      if (isFilterDefined(filter)) {
+        return this.labelValuesQuery(label, filter);
       } else {
-        return this.labelValuesQuery(labelValuesQuery[2]);
+        // Exclude the filter part of the expression because it is blank or empty
+        return this.labelValuesQuery(label);
       }
     }
 
@@ -61,79 +83,36 @@ export default class PrometheusMetricFindQuery {
 
     const queryResultQuery = this.query.match(queryResultRegex);
     if (queryResultQuery) {
-      return lastValueFrom(this.queryResultQuery(queryResultQuery[1]));
+      return this.queryResultQuery(queryResultQuery[1]);
     }
 
     // if query contains full metric name, return metric name and label list
-    return this.metricNameAndLabelsQuery(this.query);
-  }
+    const expressions = ['label_values()', 'metrics()', 'query_result()'];
+    if (!expressions.includes(this.query)) {
+      return this.metricNameAndLabelsQuery(this.query);
+    }
 
-  labelNamesQuery() {
-    const start = this.datasource.getPrometheusTime(this.range.from, false);
-    const end = this.datasource.getPrometheusTime(this.range.to, true);
-    const limit = this.datasource.getLimitMetrics('maxTagKeys');
-    const params = {
-      start: start.toString(),
-      end: end.toString(),
-      limit,
-    };
-
-    const url = `/api/v1/labels`;
-
-    return this.datasource.metadataRequest(url, params).then((result: any) => {
-      return _map(result.data.data, (value) => {
-        return { text: value };
-      });
-    });
+    return Promise.resolve([]);
   }
 
   labelValuesQuery(label: string, metric?: string) {
-    const start = this.datasource.getPrometheusTime(this.range.from, false);
-    const end = this.datasource.getPrometheusTime(this.range.to, true);
-    const limit = this.datasource.getLimitMetrics(!metric ? 'maxTagValues' : 'maxSeries')
-    const params = {
-      ...(metric && { 'match[]': metric }),
-      start: start.toString(),
-      end: end.toString(),
-      limit
-    };
+    const start = getVictoriaMetricsTime(this.range.from, false);
+    const end = getVictoriaMetricsTime(this.range.to, true);
+    const params = { ...(metric && { 'match[]': metric }), start: start.toString(), end: end.toString() };
 
-    if (!metric) {
-      const url = `/api/v1/label/${label}/values`;
+    const url = `/api/v1/label/${label}/values`;
 
-      return this.datasource.metadataRequest(url, params).then((result: any) => {
-        return _map(result.data.data, (value) => {
-          return { text: value };
-        });
-      });
-    } else {
-      const url = `/api/v1/series`;
-
-      return this.datasource.metadataRequest(url, params).then((result: any) => {
-        const _labels = _map(result.data.data, (metric) => {
-          return metric[label] || '';
-        }).filter((label) => {
-          return label !== '';
-        });
-
-        return uniq(_labels).map((metric) => {
-          return {
-            text: metric,
-            expandable: true,
-          };
-        });
-      });
-    }
+    return this.datasource.metadataRequest(url, params).then((result: any) => {
+      return _map(result.data.data, (value) => ({ text: value }));
+    });
   }
 
   metricNameQuery(metricFilterPattern: string) {
-    const start = this.datasource.getPrometheusTime(this.range.from, false);
-    const end = this.datasource.getPrometheusTime(this.range.to, true);
-    const limit = this.datasource.getLimitMetrics('maxTagValues');
+    const start = getVictoriaMetricsTime(this.range.from, false);
+    const end = getVictoriaMetricsTime(this.range.to, true);
     const params = {
       start: start.toString(),
       end: end.toString(),
-      limit,
     };
     const url = `/api/v1/label/__name__/values`;
 
@@ -154,52 +133,51 @@ export default class PrometheusMetricFindQuery {
   }
 
   queryResultQuery(query: string) {
-    const end = this.datasource.getPrometheusTime(this.range.to, true);
-    const instantQuery: PromQueryRequest = { expr: query } as PromQueryRequest;
-    return this.datasource.performInstantQuery(instantQuery, end).pipe(
-      map((result) => {
-        switch (result.data.data.resultType) {
-          case 'scalar': // [ <unix_time>, "<scalar_value>" ]
-          case 'string': // [ <unix_time>, "<string_value>" ]
-            return [
-              {
-                text: result.data.data.result[1] || '',
-                expandable: false,
-              },
-            ];
-          case 'vector':
-            return _map(result.data.data.result, (metricData) => {
-              let text = metricData.metric.__name__ || '';
-              delete metricData.metric.__name__;
-              text +=
-                '{' +
-                _map(metricData.metric, (v, k) => {
-                  return k + '="' + v + '"';
-                }).join(',') +
-                '}';
-              text += ' ' + metricData.value[1] + ' ' + metricData.value[0] * 1000;
+    const url = '/api/v1/query';
+    const params = {
+      query,
+      time: getVictoriaMetricsTime(this.range.to, true).toString(),
+    };
+    return this.datasource.metadataRequest(url, params).then((result: any) => {
+      switch (result.data.data.resultType) {
+        case 'scalar': // [ <unix_time>, "<scalar_value>" ]
+        case 'string': // [ <unix_time>, "<string_value>" ]
+          return [
+            {
+              text: result.data.data.result[1] || '',
+              expandable: false,
+            },
+          ];
+        case 'vector':
+          return _map(result.data.data.result, (metricData) => {
+            let text = metricData.metric.__name__ || '';
+            delete metricData.metric.__name__;
+            text +=
+              '{' +
+              _map(metricData.metric, (v, k) => {
+                return k + '="' + v + '"';
+              }).join(',') +
+              '}';
+            text += ' ' + metricData.value[1] + ' ' + metricData.value[0] * 1000;
 
-              return {
-                text: text,
-                expandable: true,
-              };
-            });
-          default:
-            throw Error(`Unknown/Unhandled result type: [${result.data.data.resultType}]`);
-        }
-      })
-    );
+            return {
+              text: text,
+              expandable: true,
+            };
+          });
+        default:
+          throw Error(`Unknown/Unhandled result type: [${result.data.data.resultType}]`);
+      }
+    });
   }
 
   metricNameAndLabelsQuery(query: string): Promise<MetricFindValue[]> {
-    const start = this.datasource.getPrometheusTime(this.range.from, false);
-    const end = this.datasource.getPrometheusTime(this.range.to, true);
-    const limit = this.datasource.getLimitMetrics('maxSeries')
+    const start = getVictoriaMetricsTime(this.range.from, false);
+    const end = getVictoriaMetricsTime(this.range.to, true);
     const params = {
       'match[]': query,
       start: start.toString(),
       end: end.toString(),
-      limit,
     };
 
     const url = `/api/v1/series`;
@@ -214,4 +192,9 @@ export default class PrometheusMetricFindQuery {
       });
     });
   }
-};
+}
+
+function isFilterDefined(filter: string) {
+  // We consider blank strings or the empty filter {} as an undefined filter
+  return filter && filter.split(' ').join('') !== '{}';
+}